+++ tests/test_django.py
@@ -10,6 +10,7 @@ from unittest import mock
 import django
 from django import test as django_test
 from django.conf import settings
+from django.contrib.auth.hashers import check_password
 from django.db.models import signals
 from django.test import utils as django_test_utils
 
@@ -95,6 +96,16 @@ class AbstractSonFactory(AbstractBaseFactory):
 class ConcreteGrandSonFactory(AbstractBaseFactory):
     class Meta:
         model = models.ConcreteGrandSon
+
+
+PASSWORD = 's0_s3cr3t'
+
+
+class WithPasswordFactory(factory.django.DjangoModelFactory):
+    class Meta:
+        model = models.WithPassword
+
+    pw = factory.django.Password(password=PASSWORD)
 
 
 class WithFileFactory(factory.django.DjangoModelFactory):
@@ -414,6 +425,9 @@ class DjangoRelatedFieldTestCase(django_test.TestCase):
                 factory_related_name='pointed',
             )
 
+            class Meta:
+                skip_postgeneration_save = True
+
         class PointerExtraFactory(PointerFactory):
             pointed__foo = 'extra_new_foo'
 
@@ -429,6 +443,9 @@ class DjangoRelatedFieldTestCase(django_test.TestCase):
                         bar='with_trait',
                     )
                 )
+
+            class Meta:
+                skip_postgeneration_save = True
 
         cls.PointedFactory = PointedFactory
         cls.PointerFactory = PointerFactory
@@ -490,6 +507,21 @@ class DjangoRelatedFieldTestCase(django_test.TestCase):
         self.assertEqual(pointed.foo, 'foo')
         self.assertEqual(pointed.pointer, models.PointerModel.objects.get())
         self.assertEqual(pointed.pointer.bar, 'with_trait')
+
+
+class DjangoPasswordTestCase(django_test.TestCase):
+    def test_build(self):
+        u = WithPasswordFactory.build()
+        self.assertTrue(check_password(PASSWORD, u.pw))
+
+    def test_build_with_kwargs(self):
+        password = 'V3R¥.S€C®€T'
+        u = WithPasswordFactory.build(pw=password)
+        self.assertTrue(check_password(password, u.pw))
+
+    def test_create(self):
+        u = WithPasswordFactory.create()
+        self.assertTrue(check_password(PASSWORD, u.pw))
 
 
 class DjangoFileFieldTestCase(django_test.TestCase):
@@ -909,6 +941,7 @@ class PreventSignalsTestCase(django_test.TestCase):
         class WithSignalsDecoratedFactory(factory.django.DjangoModelFactory):
             class Meta:
                 model = models.WithSignals
+                skip_postgeneration_save = True
 
             @factory.post_generation
             def post(obj, create, extracted, **kwargs):
@@ -995,6 +1028,7 @@ class PreventChainedSignalsTestCase(django_test.TestCase):
         class UndecoratedFactory(factory.django.DjangoModelFactory):
             class Meta:
                 model = models.PointerModel
+                skip_postgeneration_save = True
             pointed = factory.RelatedFactory(self.WithSignalsDecoratedFactory)
 
         UndecoratedFactory()
@@ -1015,3 +1049,28 @@ class DjangoCustomManagerTestCase(django_test.TestCase):
         # Our CustomManager will remove the 'arg=' argument,
         # invalid for the actual model.
         ObjFactory.create(arg='invalid')
+
+
+class DjangoModelFactoryDuplicateSaveDeprecationTest(django_test.TestCase):
+    class StandardFactoryWithPost(StandardFactory):
+        @factory.post_generation
+        def post_action(obj, create, extracted, **kwargs):
+            return 3
+
+    def test_create_warning(self):
+        with self.assertWarns(DeprecationWarning) as cm:
+            self.StandardFactoryWithPost.create()
+
+        [msg] = cm.warning.args
+        self.assertEqual(
+            msg,
+            "StandardFactoryWithPost._after_postgeneration will stop saving the "
+            "instance after postgeneration hooks in the next major release.\n"
+            "If the save call is extraneous, set skip_postgeneration_save=True in the "
+            "StandardFactoryWithPost.Meta.\n"
+            "To keep saving the instance, move the save call to your postgeneration "
+            "hooks or override _after_postgeneration.",
+        )
+
+    def test_build_no_warning(self):
+        self.StandardFactoryWithPost.build()